// src/models/People.ts
import { IGetBaseListApiResponse } from '@/models/Base';

export interface IPeople {
    name: string;
    birth_year: string;
    eye_color: string;
    gender: string;
    hair_color: string;
    height: string;
    mass: string;
    skin_color: string;
    homeworld: string;
    films: string[];
    species: string[];
    starships: string[];
    vehicles: string[];
    url: string;
}

export interface IGetPeopleApiResponse extends IGetBaseListApiResponse {
    results: IPeople[];
}