// src/models/Starship.ts
import { IGetBaseListApiResponse } from '@/models/Base';

export interface IStarship {
    name: string;
    model: string;
    starship_class: string;
    manufacturer: string;
    cost_in_credits: string;
    length: string;
    crew: string;
    passengers: string;
    max_atmosphering_speed: string;
    hyperdrive_rating: string;
    MGLT: string;
    cargo_capacity: string;
    consumables: string;
    films: string[];
    pilots: string[];
    url: string;
}

export interface IGetStarshipsApiResponse extends IGetBaseListApiResponse {
    results: IStarship[];
}